# api/openapi.yaml
openapi: 3.1.0
info:
  title: Reticora CMDB API
  version: 1.0.0
  description: |
    Multi-tenant Configuration Management Database API for MSPs and ISPs.
    Endpoints require tenant-scoped authentication unless explicitly marked otherwise.
  contact:
    name: DataHub Chiemgau
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/assets:
    get:
      operationId: listAssets
      tags: [Assets]
      summary: List assets
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/ClientID'
      responses:
        '200':
          description: Paginated list of assets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createAsset
      tags: [Assets]
      summary: Create an asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '201':
          description: Asset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/assets/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    get:
      operationId: getAsset
      tags: [Assets]
      summary: Get an asset
      responses:
        '200':
          description: Asset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateAsset
      tags: [Assets]
      summary: Update an asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssetRequest'
      responses:
        '200':
          description: Asset updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteAsset
      tags: [Assets]
      summary: Delete an asset
      responses:
        '204':
          description: Asset deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/search:
    get:
      operationId: search
      tags: [Search]
      summary: Search tenant data
      parameters:
        - name: q
          in: query
          schema: { type: string, maxLength: 512 }
        - name: type
          in: query
          schema: { type: string }
          description: Comma-separated entity type filter.
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/ai/conversations:
    get:
      operationId: listAIConversations
      tags: [AI]
      summary: List assistant conversations
      responses:
        '200':
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AIConversation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createAIConversation
      tags: [AI]
      summary: Create an assistant conversation
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
      responses:
        '201':
          description: Created conversation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AIConversation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/ai/ask:
    post:
      operationId: askAI
      tags: [AI]
      summary: Ask the governed RAG assistant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AIAskRequest' }
      responses:
        '200':
          description: Assistant answer
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AIAskResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource identifier.
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
      description: Maximum number of items to return.
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: |
        Number of items to skip before returning data. Ignored when `cursor` is
        supplied.
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: |
        Opaque keyset cursor taken from `next_cursor` of the previous page.
        Cursors are bound to the sort order they were issued for; sending one
        with a different `sort_by`/`sort_dir` combination yields 400.
    Search:
      name: search
      in: query
      schema:
        type: string
      description: Free-text search term.
    SortBy:
      name: sort_by
      in: query
      schema:
        type: string
      description: Sort field name.
    SortDir:
      name: sort_dir
      in: query
      schema:
        type: string
        enum: [asc, desc]
      description: Sort direction.
...
    ClientID:
      name: client_id
      in: query
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    ServiceUnavailable:
      description: A required subsystem is not configured or temporarily unavailable
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    InternalServerError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
...
    ProblemDetail:
      type: object
      description: RFC 7807 Problem Details object.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        trace_id:
          type: string
    PaginationEnvelope:
      type: object
      required: [total, limit, offset, has_more]
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
        next_cursor:
          type: string
          description: |
            Opaque cursor for the next page. Absent when no further rows follow.
...
    Asset:
      type: object
      required: [id, organization_id, asset_tag, name, category, status, custom_fields, created_at, updated_at]
      properties:
        id:
          type: string
        organization_id:
          type: string
        client_id:
          type: string
        ci_id:
          type: string
        asset_tag:
          type: string
        name:
          type: string
        category:
          type: string
        status:
          type: string
        purchase_date:
          type: string
          format: date
        purchase_cost:
          type: number
        currency:
          type: string
        warranty_end:
          type: string
          format: date
        supplier:
          type: string
        invoice_number:
          type: string
        serial_number:
          type: string
        location:
          type: string
        notes:
          type: string
        custom_fields:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateAssetRequest:
      type: object
      required: [asset_tag, name]
      properties:
        client_id:
          type: string
        ci_id:
          type: string
        asset_tag:
          type: string
        name:
          type: string
        category:
          type: string
        status:
          type: string
        purchase_date:
          type: string
          format: date
        purchase_cost:
          type: number
        currency:
          type: string
        warranty_end:
          type: string
          format: date
        supplier:
          type: string
        invoice_number:
          type: string
        serial_number:
          type: string
        location:
          type: string
        notes:
          type: string
        custom_fields:
          type: object
          additionalProperties: true
    UpdateAssetRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        status:
          type: string
        purchase_date:
          type: string
          format: date
        purchase_cost:
          type: number
        currency:
          type: string
        warranty_end:
          type: string
          format: date
        supplier:
          type: string
        invoice_number:
          type: string
        serial_number:
          type: string
        location:
          type: string
        notes:
          type: string
        ci_id:
          type: string
        custom_fields:
          type: object
          additionalProperties: true
    AssetListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Asset'
...
    SearchHit:
      type: object
      required: [id, organization_id, entity_type, entity_id, title, url, score, updated_at]
      properties:
        id: { type: string }
        organization_id: { type: string }
        entity_type: { type: string, enum: [ci, asset, document, ticket, contact, compliance] }
        entity_id: { type: string }
        title: { type: string }
        summary: { type: string }
        url: { type: string }
        score: { type: number }
        highlights:
          type: array
          items: { type: string }
        metadata:
          type: object
          additionalProperties: { type: string }
        updated_at: { type: string, format: date-time }
    SearchResponse:
      type: object
      required: [data, total, limit, offset, has_more]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/SearchHit' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        has_more: { type: boolean }
...
    AIConversation:
      type: object
      required: [id, organization_id, title, created_at, updated_at]
      properties:
        id: { type: string }
        organization_id: { type: string }
        user_id: { type: string }
        title: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AICitation:
      type: object
      required: [entity_type, entity_id, title, url, score]
      properties:
        entity_type: { type: string }
        entity_id: { type: string }
        title: { type: string }
        url: { type: string }
        score: { type: number }
    AIAskRequest:
      type: object
      required: [question]
      properties:
        conversation_id: { type: string }
        question: { type: string, maxLength: 2000 }
    AIAskResponse:
      type: object
      required: [conversation_id, answer, citations, prompt_tokens, completion_tokens]
      properties:
        conversation_id: { type: string }
        answer: { type: string }
        citations:
          type: array
          items: { $ref: '#/components/schemas/AICitation' }
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }